fix!: use impersonated service account by default for all resources
BREAKING CHANGE:
Before this change, most of the resources were not using the credentials from the APPLICATION_DEFAULT_CREDENTIALS environmental variable (which is set by running `gcloud auth application-default login`). They were still using the credentials from the logged in user instead. This can be breaking in case you did not give the required roles/permissions to the service account that is  (should) be responsible for managing your infrastructure.